Add package mlmpfr.4.0.0 (#12913)
* Add package mlmpfr.4.0.0

This commit introduces mlmpfr 4.0.0, some OCaml bindings for MPFR 4.0.0.
This version is an update of the opam package mlmpfr.3.1.6 which was release for MPFR 3.1.6.

The new features of mlmpfr 4.0.0 are:
- root function replaced by rootn_ui
- support failthful rounding mode support
- support for mpfr_flags_* functions
- support mpfr_beta
- support mpfr_fmodquo
- support mpfr_fmma and mpfr_fmms
- support mpfr_log_ui and mpfr_gamma_inc
- support mpfr_rint_roundeven and mpfr_roundeven

See https://github.com/thvnx/mlmpfr/tree/release_400.

* Add extra-files opam directive

mlmpfr will now be installed only if the right version of mpfr is installed on the system.

* Fix build test.c and add post-messages

* Fix build test

* Fix build test again

* mlmpfr.4.0.0